[common] refactoring
* 파일 위치 변경
* 필요없는 파일 제거
